Saner JS
When using `Function.prototype.call`, you're binding the function to the same
context as the caller, there's no need to name an argument to a reserved word
:trollface:.

I also huged the function definition with `()` since we don't know where this
code is going to be. It could be after a expression.

```javascript
42;function(){console.log("foo");}.call(this); # doesn't work
42;(function(){console.log("foo");}).call(this); # does!
```